refactor: improve suggest UI for groups
