+++ src/main.ts
@@ -2,8 +2,8 @@ import { MarkdownView, Plugin } from "obsidian";
 import { EnhancedEditor } from "./data/models/Settings";
 import CommanderSettingTab from "@/ui/settingTab";
 import SettingTabModal from "@/ui/modals/SettingTabModal";
-import { SlashSuggester } from "@/services/suggesters/slashSuggest";
-import { MenuSuggest } from "@/services/suggesters/menuSuggest";
+import { SlashSuggester } from "@/services/suggest/slashSuggest";
+import { MenuSuggest } from "@/services/suggest/menuSuggest";
 
 import "@/ui/styles/styles.scss";
 import registerCustomIcons from "@/assets/icons";
